Move settings test code around
