+++ tests/config/test_download_settings.py
@@ -109,11 +109,11 @@ def test_download_settings_destination_test_buttons_exist():
 
 
 def test_test_books_destination_uses_current_values(tmp_path):
-    from shelfmark.config.settings import _test_books_destination
+    from shelfmark.config.download_settings_handlers import test_books_destination
 
     destination = tmp_path / "books"
 
-    result = _test_books_destination({"DESTINATION": str(destination)})
+    result = test_books_destination({"DESTINATION": str(destination)})
 
     assert result["success"] is True
     assert result["message"] == f"Books destination is writable: {destination}"
@@ -121,11 +121,11 @@ def test_test_books_destination_uses_current_values(tmp_path):
 
 
 def test_test_audiobook_destination_falls_back_to_books_destination(tmp_path):
-    from shelfmark.config.settings import _test_audiobook_destination
+    from shelfmark.config.download_settings_handlers import test_audiobook_destination
 
     destination = tmp_path / "books"
 
-    result = _test_audiobook_destination(
+    result = test_audiobook_destination(
         {
             "DESTINATION": str(destination),
             "DESTINATION_AUDIOBOOK": "",
@@ -140,11 +140,11 @@ def test_test_audiobook_destination_falls_back_to_books_destination(tmp_path):
 
 
 def test_test_books_destination_uses_base_path_for_user_placeholder(tmp_path):
-    from shelfmark.config.settings import _test_books_destination
+    from shelfmark.config.download_settings_handlers import test_books_destination
 
     destination = tmp_path / "books"
 
-    result = _test_books_destination({"DESTINATION": f"{destination}/{{User}}"})
+    result = test_books_destination({"DESTINATION": f"{destination}/{{User}}"})
 
     assert result["success"] is True
     assert result["message"] == f"Books destination is writable: {destination}"
@@ -156,9 +156,9 @@ def test_test_books_destination_uses_base_path_for_user_placeholder(tmp_path):
 
 
 def test_test_books_destination_rejects_relative_user_placeholder_path():
-    from shelfmark.config.settings import _test_books_destination
+    from shelfmark.config.download_settings_handlers import test_books_destination
 
-    result = _test_books_destination({"DESTINATION": "{User}/books"})
+    result = test_books_destination({"DESTINATION": "{User}/books"})
 
     assert result["success"] is False
     assert result["message"] == "Destination must be absolute: {User}/books"
